Finalized SpindexerColorSensor Firmware
Update Color sensor class, updated to style guide standards, renamed class to be specific for use, as it is mostly season specific.
